leam-base.joda-time.mid-2090.idx-176407.6.mutant

+++ joda-time/src/main/java/org/joda/time/field/ScaledDurationField.java
@@ -104,7 +104,7 @@ public class ScaledDurationField extends DecoratedDurationField {
     @Override
     public long add(long instant, long value) {
         long scaled = FieldUtils.safeMultiply(value, iScalar);
-        return getWrappedField().add(instant, scaled);
+        return (long)getWrappedField().add(instant, scaled);
     }
 
     @Override
